# [perf] avoid memory allocations from converting []byte to string