--- a/gateway/analytics.go
+++ b/gateway/analytics.go
@@ -179,7 +179,7 @@ func (a *AnalyticsRecord) SetExpiry(expiresInSeconds int64) {
 // RedisAnalyticsHandler will record analytics data to a redis back end
 // as defined in the Config object
 type RedisAnalyticsHandler struct {
-	Store            storage.Handler
+	Store            storage.AnalyticsHandler
 	Clean            Purger
 	GeoIPDB          *maxminddb.Reader
 	globalConf       config.Config
@@ -257,7 +257,7 @@ func (r *RedisAnalyticsHandler) recordWorker() {
 
 	// this is buffer to send one pipelined command to redis
 	// use r.recordsBufferSize as cap to reduce slice re-allocations
-	recordsBuffer := make([]string, 0, r.workerBufferSize)
+	recordsBuffer := make([][]byte, 0, r.workerBufferSize)
 
 	// read records from channel and process
 	lastSentTs := time.Now()
@@ -269,7 +269,7 @@ func (r *RedisAnalyticsHandler) recordWorker() {
 			// check if channel was closed and it is time to exit from worker
 			if !ok {
 				// send what is left in buffer
-				r.Store.AppendToSetPipelined(analyticsKeyName, recordsBuffer)
+				r.Store.AppendToSetPipelinedBytes(analyticsKeyName, recordsBuffer)
 				return
 			}
 
@@ -310,7 +310,7 @@ func (r *RedisAnalyticsHandler) recordWorker() {
 			if encoded, err := msgpack.Marshal(record); err != nil {
 				log.WithError(err).Error("Error encoding analytics data")
 			} else {
-				recordsBuffer = append(recordsBuffer, string(encoded))
+				recordsBuffer = append(recordsBuffer, encoded)
 			}
 
 			// identify that buffer is ready to be sent
@@ -324,8 +324,8 @@ func (r *RedisAnalyticsHandler) recordWorker() {
 
 		// send data to Redis and reset buffer
 		if len(recordsBuffer) > 0 && (readyToSend || time.Since(lastSentTs) >= recordsBufferForcedFlushInterval) {
-			r.Store.AppendToSetPipelined(analyticsKeyName, recordsBuffer)
-			recordsBuffer = make([]string, 0, r.workerBufferSize)
+			r.Store.AppendToSetPipelinedBytes(analyticsKeyName, recordsBuffer)
+			recordsBuffer = recordsBuffer[:0]
 			lastSentTs = time.Now()
 		}
 	}