--- a/main.go
+++ b/main.go
@@ -1,6 +1,8 @@
 package main
 
-import "github.com/TykTechnologies/tyk/gateway"
+import (
+	"github.com/TykTechnologies/tyk/gateway"
+)
 
 func main() {
 	gateway.Start()