@buger 2019-09-06
Why you added new interface instead of changing storage.Handler ?