@asoorm 2019-11-01
> Why you added new interface instead of changing storage.Handler ?

The bigger the interface, the weaker the abstraction