@dencoded 2019-11-03
I think it should log less info in general, only useful info like warnings in this PR. So thats minify logrus overhead.

According to memory optimizations - formatted output into string might be heavier on CPU or RAM.